🧪 Ejercicio 5 - Composición
This commit introduces a new `Direccion` class and updates the `Cliente` class to include an address. This allows us to store and display the address of a client. Getters and setters were added to the Cliente class to allow modification of the attributes. A toString method was added to the Cliente class to display the attributes of the class.